% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deng_obj.R
\name{Deng}
\alias{Deng}
\title{deng_obj  data engineering object (deng)}
\description{
deng_obj  data engineering object (deng)

deng_obj  data engineering object (deng)
}
\details{
An object that transforms data into more robust forms.
This includes pairs, signature-pairs, quartiles, and ranks
}
\examples{
# New object
ann <- Recm$new("Ann")

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{what}}{type of engineering to do}

\item{\code{if}}{there are signatures to be used}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{what}}{type of engineering to do}

\item{\code{if}}{there are signatures to be used}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Deng$new()}}
\item \href{#method-data_eng}{\code{Deng$data_eng()}}
\item \href{#method-clone}{\code{Deng$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Recm` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deng$new(data_mode = NULL, signatures = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The object is named.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `recm` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_eng"></a>}}
\if{latex}{\out{\hypertarget{method-data_eng}{}}}
\subsection{Method \code{data_eng()}}{
Data engineering, replaces the object's data.table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deng$data_eng(data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Deng$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
