% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enbl_obj.R
\name{Ensbl}
\alias{Ensbl}
\title{An ensemble of xgboosts}
\description{
Each member of the ensemble is an xgboost, and contains a random sample of the data.
}
\details{
The ensbl object contains a list of xgboost objects
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Ensbl$new()}}
\item \href{#method-print}{\code{Ensbl$print()}}
\item \href{#method-sample_data}{\code{Ensbl$sample_data()}}
\item \href{#method-train_models}{\code{Ensbl$train_models()}}
\item \href{#method-ensemble_combine}{\code{Ensbl$ensemble_combine()}}
\item \href{#method-final_ensemble_combine}{\code{Ensbl$final_ensemble_combine()}}
\item \href{#method-member_predict}{\code{Ensbl$member_predict()}}
\item \href{#method-print_error}{\code{Ensbl$print_error()}}
\item \href{#method-clone}{\code{Ensbl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$new(name, obj_mode, size, data, label, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_data"></a>}}
\if{latex}{\out{\hypertarget{method-sample_data}{}}}
\subsection{Method \code{sample_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$sample_data(perc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_models"></a>}}
\if{latex}{\out{\hypertarget{method-train_models}{}}}
\subsection{Method \code{train_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$train_models(perc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ensemble_combine"></a>}}
\if{latex}{\out{\hypertarget{method-ensemble_combine}{}}}
\subsection{Method \code{ensemble_combine()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$ensemble_combine(combine_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-final_ensemble_combine"></a>}}
\if{latex}{\out{\hypertarget{method-final_ensemble_combine}{}}}
\subsection{Method \code{final_ensemble_combine()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$final_ensemble_combine(combine_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-member_predict"></a>}}
\if{latex}{\out{\hypertarget{method-member_predict}{}}}
\subsection{Method \code{member_predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$member_predict(data, combine_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_error"></a>}}
\if{latex}{\out{\hypertarget{method-print_error}{}}}
\subsection{Method \code{print_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$print_error(label, threshold)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Ensbl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
