% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recm_obj.R
\name{Recm}
\alias{Recm}
\title{Recm  Robust ensemble classifier machine (Recm)}
\description{
Recm  Robust ensemble classifier machine (Recm)

Recm  Robust ensemble classifier machine (Recm)
}
\details{
An object that holds data and an ensemble of classifiers
The ensemble is composed of XGBoost classifiers trained on binarized labels.

A list of classifiers, each trained on a random sample of the training data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{the object's name}

\item{\code{data_mode}}{string vector, dictates the types of data engineeering, acceptable values include combination of: original  quartiles  pairs  ranks  sigpairs}

\item{\code{signatures}}{lists of variables, must be in data, that will be used together, and compared to other signatures}

\item{\code{file_name}}{the data file}

\item{\code{data}}{the data.table used to train and test}

\item{\code{the}}{data column used as label, the target}

\item{\code{data_split}}{numeric value indicating percent data to make into training data}

\item{\code{train_data}}{the data.table used to train the ensemble}

\item{\code{train_label}}{the vector used as the training label}

\item{\code{test_data}}{the data.table used as test data}

\item{\code{test_label}}{the vector used as test data labels}

\item{\code{data_colnames}}{the column names used to train the models}

\item{\code{unique_labels}}{the unique set of labels}

\item{\code{the}}{ensemble of predictors}

\item{\code{the}}{table of predictions, collecting from the ensbl list of predictors}

\item{\code{the}}{table of predictions, collecting from the ensbl list of predictors}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{the}}{data column used as label, the target}

\item{\code{the}}{ensemble of predictors}

\item{\code{the}}{table of predictions, collecting from the ensbl list of predictors}

\item{\code{the}}{table of predictions, collecting from the ensbl list of predictors}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Recm$new()}}
\item \href{#method-greet}{\code{Recm$greet()}}
\item \href{#method-read_data}{\code{Recm$read_data()}}
\item \href{#method-read_train_data}{\code{Recm$read_train_data()}}
\item \href{#method-read_test_data}{\code{Recm$read_test_data()}}
\item \href{#method-data_eng}{\code{Recm$data_eng()}}
\item \href{#method-data_setup}{\code{Recm$data_setup()}}
\item \href{#method-train_data_setup}{\code{Recm$train_data_setup()}}
\item \href{#method-test_data_setup}{\code{Recm$test_data_setup()}}
\item \href{#method-binarize_label}{\code{Recm$binarize_label()}}
\item \href{#method-build_label_ensemble}{\code{Recm$build_label_ensemble()}}
\item \href{#method-build_pred_table}{\code{Recm$build_pred_table()}}
\item \href{#method-remap_multiclass_labels}{\code{Recm$remap_multiclass_labels()}}
\item \href{#method-unmap_multiclass_labels}{\code{Recm$unmap_multiclass_labels()}}
\item \href{#method-build_final_ensemble}{\code{Recm$build_final_ensemble()}}
\item \href{#method-train_models}{\code{Recm$train_models()}}
\item \href{#method-train_final}{\code{Recm$train_final()}}
\item \href{#method-ensemble_predict}{\code{Recm$ensemble_predict()}}
\item \href{#method-predict}{\code{Recm$predict()}}
\item \href{#method-print_error}{\code{Recm$print_error()}}
\item \href{#method-accuracy}{\code{Recm$accuracy()}}
\item \href{#method-precision}{\code{Recm$precision()}}
\item \href{#method-sensitivity}{\code{Recm$sensitivity()}}
\item \href{#method-specificity}{\code{Recm$specificity()}}
\item \href{#method-classification_metrics}{\code{Recm$classification_metrics()}}
\item \href{#method-importance}{\code{Recm$importance()}}
\item \href{#method-results}{\code{Recm$results()}}
\item \href{#method-autopred}{\code{Recm$autopred()}}
\item \href{#method-clone}{\code{Recm$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Recm` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$new(name = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The object is named.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `recm` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-greet"></a>}}
\if{latex}{\out{\hypertarget{method-greet}{}}}
\subsection{Method \code{greet()}}{
Creates a printable representation of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$greet()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character string representing the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_data"></a>}}
\if{latex}{\out{\hypertarget{method-read_data}{}}}
\subsection{Method \code{read_data()}}{
Reads the data file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$read_data(file_name, sep, header)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{The name of the file.}

\item{\code{sep}}{The separting character ',' or '\t'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_train_data"></a>}}
\if{latex}{\out{\hypertarget{method-read_train_data}{}}}
\subsection{Method \code{read_train_data()}}{
Reads the data file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$read_train_data(file_name, sep, header)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{The name of the file.}

\item{\code{sep}}{The separting character ',' or '\t'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_test_data"></a>}}
\if{latex}{\out{\hypertarget{method-read_test_data}{}}}
\subsection{Method \code{read_test_data()}}{
Reads the data file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$read_test_data(file_name, sep, header)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{The name of the file.}

\item{\code{sep}}{The separting character ',' or '\t'}

\item{\code{header}}{boolean, whether the table has a header line}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_eng"></a>}}
\if{latex}{\out{\hypertarget{method-data_eng}{}}}
\subsection{Method \code{data_eng()}}{
Data engineering, replaces the object's data.table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$data_eng(data_source = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_setup"></a>}}
\if{latex}{\out{\hypertarget{method-data_setup}{}}}
\subsection{Method \code{data_setup()}}{
Does some setup processing on the data file, drop columns, split data into train and test, and identify the label column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$data_setup(
  file_name = NULL,
  sep = NULL,
  data_mode = NULL,
  signatures = NULL,
  label_name = NULL,
  sample_id = NULL,
  drop_list = NULL,
  data_split = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{string, the name of the file}

\item{\code{sep}}{string, the separating character}

\item{\code{data_mode}}{string,}

\item{\code{label_name}}{string, the column name indicating the target label}

\item{\code{drop_list}}{a vector of strings indicating what columns to drop}

\item{\code{data_split}}{numeric value, the percent of data to use in training}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_data_setup"></a>}}
\if{latex}{\out{\hypertarget{method-train_data_setup}{}}}
\subsection{Method \code{train_data_setup()}}{
Does some setup processing on the training data file, drop columns and identify the label column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$train_data_setup(
  file_name = NULL,
  sep = NULL,
  data_mode = NULL,
  signatures = NULL,
  label_name = NULL,
  sample_id = NULL,
  drop_list = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label_name}}{string, the column name indicating the target label}

\item{\code{drop_list}}{a vector of strings indicating what columns to drop}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-test_data_setup"></a>}}
\if{latex}{\out{\hypertarget{method-test_data_setup}{}}}
\subsection{Method \code{test_data_setup()}}{
Does some setup processing on the test data file, drop columns and identify the label column.
The data_mode and signatures will have already been set in training.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$test_data_setup(
  file_name = NULL,
  sep = NULL,
  label_name = NULL,
  sample_id = NULL,
  drop_list = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label_name}}{string, the column name indicating the target label}

\item{\code{drop_list}}{a vector of strings indicating what columns to drop}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-binarize_label"></a>}}
\if{latex}{\out{\hypertarget{method-binarize_label}{}}}
\subsection{Method \code{binarize_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$binarize_label(label, x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_label_ensemble"></a>}}
\if{latex}{\out{\hypertarget{method-build_label_ensemble}{}}}
\subsection{Method \code{build_label_ensemble()}}{
Builds list of ensembles of XGBoost object, each classifying one binary label.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$build_label_ensemble(size, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{numeric, number of classifiers}

\item{\code{mode}}{character vector, what types of data modalities to make. possible: pairs, quartiles, set-pairs}

\item{\code{label}}{string, the label vector of each data example}

\item{\code{max_depth}}{numeric, the depth of the tree in XGBoost}

\item{\code{eta}}{numeric, the eta param of XGBoost, speed of learning}

\item{\code{nrounds}}{numeric, the number of training rounds}

\item{\code{nthreads}}{numeric, the number of threads to use in processing}

\item{\code{objective}}{string, binary:logistic, see xgboost docs}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ensemble object is added to the list of objects in recm$enbl.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_pred_table"></a>}}
\if{latex}{\out{\hypertarget{method-build_pred_table}{}}}
\subsection{Method \code{build_pred_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$build_pred_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-remap_multiclass_labels"></a>}}
\if{latex}{\out{\hypertarget{method-remap_multiclass_labels}{}}}
\subsection{Method \code{remap_multiclass_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$remap_multiclass_labels(label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unmap_multiclass_labels"></a>}}
\if{latex}{\out{\hypertarget{method-unmap_multiclass_labels}{}}}
\subsection{Method \code{unmap_multiclass_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$unmap_multiclass_labels(labels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-build_final_ensemble"></a>}}
\if{latex}{\out{\hypertarget{method-build_final_ensemble}{}}}
\subsection{Method \code{build_final_ensemble()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$build_final_ensemble(size, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_models"></a>}}
\if{latex}{\out{\hypertarget{method-train_models}{}}}
\subsection{Method \code{train_models()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$train_models(perc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_final"></a>}}
\if{latex}{\out{\hypertarget{method-train_final}{}}}
\subsection{Method \code{train_final()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$train_final(perc)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ensemble_predict"></a>}}
\if{latex}{\out{\hypertarget{method-ensemble_predict}{}}}
\subsection{Method \code{ensemble_predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$ensemble_predict(data, combine_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$predict(data, combine_function)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print_error"></a>}}
\if{latex}{\out{\hypertarget{method-print_error}{}}}
\subsection{Method \code{print_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$print_error(label, root, threshold)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-accuracy"></a>}}
\if{latex}{\out{\hypertarget{method-accuracy}{}}}
\subsection{Method \code{accuracy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$accuracy(labels, calls)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-precision"></a>}}
\if{latex}{\out{\hypertarget{method-precision}{}}}
\subsection{Method \code{precision()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$precision(cmdf, i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sensitivity"></a>}}
\if{latex}{\out{\hypertarget{method-sensitivity}{}}}
\subsection{Method \code{sensitivity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$sensitivity(cmdf, i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-specificity"></a>}}
\if{latex}{\out{\hypertarget{method-specificity}{}}}
\subsection{Method \code{specificity()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$specificity(cmdf, i)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-classification_metrics"></a>}}
\if{latex}{\out{\hypertarget{method-classification_metrics}{}}}
\subsection{Method \code{classification_metrics()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$classification_metrics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance"></a>}}
\if{latex}{\out{\hypertarget{method-importance}{}}}
\subsection{Method \code{importance()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$importance()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-results"></a>}}
\if{latex}{\out{\hypertarget{method-results}{}}}
\subsection{Method \code{results()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$results(include_label = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-autopred"></a>}}
\if{latex}{\out{\hypertarget{method-autopred}{}}}
\subsection{Method \code{autopred()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$autopred(
  data_file = NULL,
  sep = NULL,
  label_name = NULL,
  sample_id = NULL,
  drop_list = NULL,
  data_split = NULL,
  data_mode = NULL,
  signatures = NULL,
  size = NULL,
  params = NULL,
  train_perc = NULL,
  combine_function = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recm$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
